refactor: componentize turn results section
