+++ src/app/page.tsx
@@ -1,5 +1,6 @@
 "use client";
 import Sketchpad from "@/components/Sketchpad";
+import TurnResultSection from "@/components/TurnResultSection";
 import { GuessState, SketchpadRef, TurnCycleState } from "@/utils/types";
 import { useEffect, useState, useRef } from "react";
 import { getRandomPrompt } from "@/utils/get-random-prompt";
@@ -26,13 +27,6 @@ export default function Home() {
     }
   };
 
-  // Visual Indicator for guess correctness, this is a placeholder for a more advanced scoring system
-  const borderColorMap = {
-    [GuessState.Pending]: "border-black",
-    [GuessState.Correct]: "border-emerald-500",
-    [GuessState.Incorrect]: "border-red-500",
-  };
-
   return (
     <div className="flex items-center justify-center flex-col gap-4 container mx-auto">
       <h1 className="text-4xl">AI Pictionary</h1>
@@ -43,15 +37,11 @@ export default function Home() {
         setGuessState={setGuessState}
         currentDrawingPrompt={currentDrawingPrompt}
       ></Sketchpad>
-      <div className={`border-2 px-120 py-20 rounded-2xl mt-10 ${borderColorMap[guessState]}`}>
-        {response}
-      </div>
-      <button
-        onClick={handleNextPrompt}
-        className="bg-blue-400 text-white px-10 py-3 rounded-xl text-2xl shadow-lg hover:cursor-pointer transition hover:scale-110"
-      >
-        Next Prompt
-      </button>
+      <TurnResultSection
+        response={response}
+        guessState={guessState}
+        handleNextPrompt={handleNextPrompt}
+      ></TurnResultSection>
     </div>
   );
 }
